Add a policy relating to the creator of the Snippet

+++ app/Http/Controllers/SnippetController.php
@@ -12,7 +12,9 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\SnippetWrite;
 use App\Models\Snippet;
+use App\Models\User;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Auth;
@@ -32,7 +34,13 @@ class SnippetController extends Controller
      */
     public function index()
     {
-        return Snippet::paginate();
+        $authenticatedUser = Auth::user();
+
+        if ($authenticatedUser->can('view_all_snippets')) {
+            return Snippet::paginate();
+        }
+        
+        return Snippet::where('created_by', '=', $authenticatedUser->id)->paginate();
     }
 
     /**
@@ -40,9 +48,12 @@ class SnippetController extends Controller
      *
      * @param Snippet $snippet
      * @return Snippet
+     * @throws AuthorizationException
      */
     public function show(Snippet $snippet): Snippet
     {
+        $this->authorize('view', [ $snippet ]);
+
         return $snippet;
     }
 
@@ -51,9 +62,12 @@ class SnippetController extends Controller
      *
      * @param SnippetWrite $snippetWrite
      * @return Snippet
+     * @throws AuthorizationException
      */
     public function store(SnippetWrite $snippetWrite): Snippet
     {
+        $this->authorize('create', [ Auth::user() ]);
+
         $snippetFields = $snippetWrite->validated();
         $snippetFields['created_by'] = (Auth::user())->id;
 
@@ -69,9 +83,11 @@ class SnippetController extends Controller
      * @param SnippetWrite $snippetWrite
      * @param Snippet $snippet
      * @return Snippet
+     * @throws AuthorizationException
      */
     public function update(SnippetWrite $snippetWrite, Snippet $snippet): Snippet
     {
+        $this->authorize('create', [ $snippet ]);
         $snippet->update($snippetWrite->validated());
 
         return $snippet;
@@ -86,6 +102,7 @@ class SnippetController extends Controller
      */
     public function destroy(Snippet $snippet): JsonResponse
     {
+        $this->authorize('create', [ $snippet ]);
         $snippet->delete();
 
         return response()->json([], 204);
